backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
cryptography==45.0.5
requests==2.31.0
slowapi==0.1.9
passlib[bcrypt]==1.7.4
google-cloud-firestore==2.11.1
python-dateutil==2.8.2
pathlib==1.0.1
openai==2.14.0
PyYAML>=6.0
Pillow>=10.0.0
firebase-admin>=6.0.0
pypandoc==1.13
playwright==1.52.0
reportlab==4.2.2
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0
spacy==3.7.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
